Update and rename README.md to SortingArray.py
This is a python code for sorting n numbers of a sequence in increasing order. The code is divided into two steps. The first step is for each even indexed element (Python is index 0), swap the element and the number to its right, whenever the element is greater. The second step is for each odd indexed element, swap the element and the number to its right, whenever the element is greater. Repeat the the above two steps until no more swapping occurs.